requirements.txt
lumibot>=3.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
